modified:   init_server.py 	modified:   maintance.ipynb

+++ init_server.py
@@ -275,9 +275,9 @@ class stt_server:
 		self.conn.commit() # autocommit
 			
 	def remove_temporary_file(self):
-
-		print('removing',self.temp_file_path + self.temp_file_name)
-		os.remove(self.temp_file_path + self.temp_file_name)
+		if self.source_id == self.sources['call']:
+			print('removing',self.temp_file_path + self.temp_file_name)
+			os.remove(self.temp_file_path + self.temp_file_name)
 
 	def get_sql_complete_files(self):
